+++ src/main/resources/static/css/sales/OrderList.css
@@ -1,131 +1,44 @@
 /* 테이블*/
-.order-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-top: 20px;
-}
-.order-table th, .order-table td {
-    border: 1px solid #ddd;
-    padding: 8px;
-    text-align: left;
-}
-.order-table th {
-    background-color: #f2f2f2;
-}
-.order-table tr:hover {
-    background-color: #f1f1f1;
-    background-color: #ffdb0036; /* 이케아 노란색 */
-}
-.order-table a {
-    color: blue;
-    text-decoration: underline;
-}
-.orderList-title h3 {
-    font-size: 18px;
-    font-weight: bold;
-    color: #333333;
-    letter-spacing: -0.1px;
-    padding-bottom: 16px;
-    border-bottom: 1px solid #999999;
-    margin: 26px 0 0;
-    position: relative;
-}
-.orderList-container{
-    padding: 20px;
-    border-radius: 5px;
-}
-.filter-type{
-    position: relative;
-    align-items: center;
-    border-radius: 5px;
-    border: 1px solid rgba(36, 42, 48, .14);
-    display: inline-flex;
-    font-size: 14px;
-    padding-left: 12px;
-    color: #666;
-}
-.date-separator {
-    font-size: 1.5rem; /* 기호 크기 조절 */
-}
-.checkbox-input{
-    width: 50px;
-
-}
-.approve-button{
-
-}
-
-
+.order-table { width: 100%; border-collapse: collapse; margin-top: 20px; } 
+.order-table th, .order-table td { border: 1px solid #ddd; padding: 8px; text-align: left; } 
+.order-table th { background-color: #f2f2f2; } 
+.order-table tr:hover { background-color: #f1f1f1; background-color: #ffdb0036; /* 이케아 노란색 */}
+.order-table a { color: blue; text-decoration: underline; } 
+.orderList-title h3 { font-size: 18px; font-weight: bold; color: #333333; letter-spacing: -0.1px; padding-bottom: 16px; border-bottom: 1px solid #999999; margin: 26px 0 0; position: relative; } 
+.orderList-container { padding: 20px; border-radius: 5px; } 
+.filter-type { position: relative; align-items: center; border-radius: 5px; border: 1px solid rgba(36, 42, 48, .14); display: inline-flex; font-size: 14px; padding-left: 12px; color: #666; } 
+.date-separator { font-size: 1.5rem; /* 기호 크기 조절 */}
+.checkbox-input { width: 50px;}
+.approve-button {}
 
 /* 검색 필터*/
-.filter-section {
-    margin-bottom: 20px;
-}
-
+.filter-section { margin-bottom: 20px; } 
 .filter-section select,
 .filter-section input,
-.filter-section .search-button {
-    background-color: #f2f2f2;
-    margin-right: 10px;
-    padding: 5px;
-}
-.filter-button {
-    background-color: #007CC1;
-    color: white;
-    padding: 6px 15px;
-    margin: 5px;
-    margin-right: 10px;
-    border: none;
-    cursor: pointer;
-    border-radius: 5px;
-}
-.filter-button:hover {
-    background-color: #005a91; /* 호버 시 조금 더 어두운 색으로 변경 */
-}
+.filter-section .search-button { background-color: #f2f2f2; margin-right: 10px; padding: 5px; } 
+.filter-button { background-color: #007CC1; color: white; padding: 6px 15px; margin: 5px; margin-right: 10px; border: none; cursor: pointer; border-radius: 5px; } 
+.filter-button:hover { background-color: #005a91; /* 호버 시 조금 더 어두운 색으로 변경 */}
 
 /* 페이지네이션*/
-.pagination-buttons {
-    margin-top: 20px;
-    text-align: center;
-}
-.pagination-buttons button {
-    display: inline-block; /* 버튼들을 한 줄로 표시 */
-    padding: 5px 10px;
-    margin: 0 5px;
-    cursor: pointer;
-    background-color: #f2f2f2;
-    border: none;
-}
-.pagination-buttons button:hover {
-    background-color: #e0e0e0;
-}
-.pagination-buttons span {
-    margin: 0 10px;
-}
+.pagination-buttons { margin-top: 20px; text-align: center; } 
+.pagination-buttons button { display: inline-block; /* 버튼들을 한 줄로 표시 */
+ padding: 5px 10px; margin: 0 5px; cursor: pointer; background-color: #f2f2f2; border: none; } 
+.pagination-buttons button:hover { background-color: #e0e0e0; } 
+.pagination-buttons span { margin: 0 10px; } 
 
 /*로딩 중 스피너 */
-
-.spinner {
-    border: 4px solid rgba(0, 0, 0, 0.1); /* 배경의 투명한 테두리 */
-    border-left: 4px solid #3498db; /* 애니메이션의 색상 */
-    border-radius: 50%; /* 원형으로 만들기 */
-    width: 40px;
-    height: 40px;
-    animation: spin 1s linear infinite; /* 애니메이션 설정 */
-    margin: auto; /* 가운데 정렬 */
-}
+.spinner { border: 4px solid rgba(0, 0, 0, 0.1); /* 배경의 투명한 테두리 */
+ border-left: 4px solid #3498db; /* 애니메이션의 색상 */
+ border-radius: 50%; /* 원형으로 만들기 */
+ width: 40px; height: 40px; animation: spin 1s linear infinite; /* 애니메이션 설정 */
+ margin: auto; /* 가운데 정렬 */}
 
 /* 스피너 애니메이션 */
-@keyframes spin {
-    0% { transform: rotate(0deg); }
-    100% { transform: rotate(360deg); }
-}
+@keyframes spin { 
+ 0% { transform: rotate(0deg); } 
+ 100% { transform: rotate(360deg); } 
+ }
 
 /* 로딩 화면 스타일 */
-.loading-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh; /* 화면 전체 높이 */
-    background-color: #f9f9f9; /* 배경색 */
-}
+.loading-container { display: flex; justify-content: center; align-items: center; height: 100vh; /* 화면 전체 높이 */
+ background-color: #f9f9f9; /* 배경색 */}
